+++ NEWS.rst
@@ -1,3 +1,14 @@
+Changes in 3.13 (released 10/21/2010)
+-------------------------------------
+
+*	The tag the :mod:`ll.sisyphus` itself uses for logging have changed slightly.
+	For more info see the module documentation.
+
+*	:mod:`ll.sisyphus` job can now log to ``stdout`` and ``stderr``.
+
+*	The option :option:`-l` for sisyphus jobs has been renamed to :option`-f`.
+
+
 Changes in 3.12.1 (released 10/21/2010)
 ---------------------------------------